+++ docs/changelog.rst
@@ -6,6 +6,7 @@ Changelog
 Changes in dev
 ==============
 
+- Improved efficiency of .get()
 - Updated ComplexFields so if required they won't accept empty lists / dicts
 - Added spec file for rpm-based distributions
 - Fixed ListField so it doesnt accept strings